+++ src/fiber.rs
@@ -16,7 +16,6 @@ use crate::r_hash::RHash;
 use crate::{
     api::Ruby,
     block::Proc,
-    data_type_builder,
     error::{Error, protect},
     exception::Exception,
     gc,
@@ -25,7 +24,7 @@ use crate::{
     object::Object,
     r_typed_data::RTypedData,
     try_convert::TryConvert,
-    typed_data::{DataType, DataTypeFunctions},
+    typed_data::{DataType, DataTypeBuilder, DataTypeFunctions},
     value::{
         QUNDEF, ReprValue, Value,
         private::{self, ReprValue as _},
@@ -579,7 +578,7 @@ where
         }
     }
 
-    let data_type = data_type_builder!(Closure<F>, "rust closure")
+    let data_type = DataTypeBuilder::<Closure<F>>::new(c"rust closure")
         .free_immediately()
         .mark()
         .build();